+++ tests/test_exec.py
@@ -1,6 +1,6 @@
-from virtool_workflow import execute
+from virtool_workflow.execution import run_in_executor
 
 
 async def test_sh():
-    out, _ = await execute.run_shell_command(["ls"])
+    out, _ = await run_in_executor.run_shell_command(["ls"])
     assert out
